msm: We are using the msm2 gpu driver, not msm
Fix references that are incorrectly pointing at drivers/gpu/msm
instead of msm2.

The incorrect includes cause linaro gcc 4.8 with -O3 to produce
an unbootable kernel on the DLX.

While it doesn't seem to be affecting our normal builds, not
fixing it will likely bite us in the future in some other
mysterious way.

Change-Id: I3c51062ca6a66ddb9782365d2c51c5e485556c1b
Signed-off-by: Paul Keith <javelinanddart@gmail.com>

Conflicts:

	drivers/gpu/Makefile

Signed-off-by: Paul Keith <javelinanddart@gmail.com>